requirements.txt
Django>=4
django-extensions
djangorestframework
sorl-thumbnail
python-dotenv
django-import-export
humanize
django-seed
django-filter
dj-rest-auth
django-allauth
django-cors-headers
pyotp
dotty-dict
phonenumbers
djangorestframework-guardian
psycopg2-binary
django-request-logging
django-reset-migrations
pandas
django-restql
django-admin-interface
django-notifications-hq
django-rest-swagger